Initial script for 18.x
Thanks to fzinken on discourse.osmc.tv, for the ground work. Just took his post and made it into an addon. fzinken changed the password to Test2 inhis post, but a thought a "restore" to factory default would be the obvious choice for a addon. Might develop this to " chose" password in the future.